Move RainMachine utils to the correct location
