+++ homeassistant/components/rainmachine/util.py
@@ -3,6 +3,23 @@ from __future__ import annotations
 
 from typing import Any
 
+from homeassistant.backports.enum import StrEnum
+
+
+class RunStates(StrEnum):
+    """Define an enum for program/zone run states."""
+
+    NOT_RUNNING = "Not Running"
+    QUEUED = "Queued"
+    RUNNING = "Running"
+
+
+RUN_STATE_MAP = {
+    0: RunStates.NOT_RUNNING,
+    1: RunStates.RUNNING,
+    2: RunStates.QUEUED,
+}
+
 
 def key_exists(data: dict[str, Any], search_key: str) -> bool:
     """Return whether a key exists in a nested dict."""
